Moved files to new packages and added new unit tests
